+++ tests/integration/api-validation/test-data/posts.php
@@ -92,6 +92,8 @@ function populate_post( $post_id, $model, $test_class ) {
 	update_post_meta( $post_id, 'booleanRequired', 'true' );
 	update_post_meta( $post_id, 'multiSingle', [ 'kiwi' ] );
 	update_post_meta( $post_id, 'multipleMulti', [ 'apple', 'banana' ] );
+	update_post_meta( $post_id, 'featured', $ids[ $model . '_image_id' ] );
+	update_post_meta( $post_id, '_thumbnail_id', $ids[ $model . '_image_id' ] );
 
 	$image_meta = array(
 		'width'  => 1000,